Refactor application structure into modules for improved organization
