+++ src/config.rs
@@ -1,0 +1,4 @@
+pub const TERMINAL_EMULATOR: &str = "kitty";
+pub const WINDOW_WIDTH: i32 = 600;
+pub const WINDOW_HEIGHT: i32 = 260;
+pub const MAX_RESULTS: usize = 5;
